requirements.txt
flask
python-dotenv
requests
pillow
gradio_client
diffusers
torch
